# Conflicts:
#	src/app/breadcrumbs/breadcrumbs.component.html
#	src/app/breadcrumbs/breadcrumbs.component.ts
#	src/app/core/locale/locale.service.spec.ts
#	src/app/core/locale/locale.service.ts
#	src/app/cris-layout/cris-layout-matrix/cris-layout-box-container/boxes/metadata/rendering-types/longtext/longtext.component.ts
#	src/app/info/end-user-agreement/end-user-agreement-content/end-user-agreement-content.component.ts
#	src/app/shared/object-list/my-dspace-result-list-element/item-list-preview/item-list-preview.component.html
#	src/app/shared/object-list/my-dspace-result-list-element/item-list-preview/item-list-preview.component.ts
#	src/app/shared/object-list/search-result-list-element/item-search-result/item-types/item/item-search-result-list-element.component.html
#	src/app/shared/object-list/search-result-list-element/item-search-result/item-types/item/item-search-result-list-element.component.ts
#	src/themes/custom/app/breadcrumbs/breadcrumbs.component.ts
#	src/themes/custom/app/shared/object-list/my-dspace-result-list-element/item-list-preview/item-list-preview.component.ts
#	src/themes/custom/app/shared/object-list/search-result-list-element/item-search-result/item-types/item/item-search-result-list-element.component.ts

+++ src/app/core/locale/locale.service.ts
@@ -1,8 +1,10 @@
+import { isPlatformBrowser } from '@angular/common';
 import {
   DOCUMENT,
   Inject,
   Injectable,
   OnDestroy,
+  PLATFORM_ID,
 } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import {
@@ -63,6 +65,7 @@ export class LocaleService implements OnDestroy {
     protected authService: AuthService,
     protected routeService: RouteService,
     @Inject(DOCUMENT) protected document: any,
+    @Inject(PLATFORM_ID) protected platformId: string,
   ) {
     this.initDefaults();
   }
@@ -74,7 +77,9 @@ export class LocaleService implements OnDestroy {
     this.routeService.getQueryParameterValue('lang').subscribe(lang => {
       if (lang && this.translate.getLangs().includes(lang)) {
         this.setCurrentLanguageCode(lang);
-        this.routeService.removeQueryParam('lang');
+        if (isPlatformBrowser(this.platformId)) {
+          this.routeService.removeQueryParam('lang');
+        }
       }
     });
   }
